backend/requirements.txt
dotenv==0.9.9
fastapi==0.135.1
google_api_python_client==2.192.0
google_auth_oauthlib==1.3.0
googleapis-common-protos==1.73.0
httpx==0.28.1
playwright==1.58.0
protobuf>=3.20.2,<6.0.0
pydantic==2.12.5
pytest==9.0.2
pytest_asyncio==1.3.0
python-dotenv==1.2.2
python_dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
supabase==2.28.2
uvicorn==0.42.0